Added ability to only find playlist items from X year
I decided a good use case for this script would be to only scrape songs from the current year (2023) so that we can create a Spotify playlist that contains only songs that KEXP played that are from the current year... in other words, new songs only. This allows the playlist to be a tool for people to keep up to date on new music!

This works by checking each playlist item for the div element that contains the year the song was published, then we use a regular expression to see if the year is equal to the year entered into the script.

TODO: However I think it'd be cool to modify the script so that this feature is easily toggled on/off. That way I can quickly switch between scraping all songs OR scraping songs from a specific year without editing the code.