Check if Track exists in Playlist as they are added, rather than after
Major refactor which removed the code that checks for duplicate Tracks, removes them all, and then adds one instance of the Track back into the Playlist. This is not efficient and a waste of resources. Moving forward the script will simply check if the Track exists before it attempts to add it to the Playlist. If it does, then it will skip. If it does not then it will proceed with adding it.